# labels: further optimize IPStringToLabel for single IP case