--- a/pkg/labels/cidr.go
+++ b/pkg/labels/cidr.go
@@ -21,9 +21,7 @@ import (
 //
 // For IPv6 addresses, it converts ":" into "-" as EndpointSelectors don't
 // support colons inside the name section of a label.
-func maskedIPToLabel(ip netip.Addr, prefix int) Label {
-	ipStr := ip.String()
-
+func maskedIPToLabel(ipStr string, prefix int) Label {
 	var str strings.Builder
 	str.Grow(
 		1 /* preZero */ +
@@ -70,13 +68,13 @@ func IPStringToLabel(ip string) (Label, error) {
 		if err != nil {
 			return Label{}, fmt.Errorf("%q is not an IP address: %w", ip, err)
 		}
-		return maskedIPToLabel(parsedIP, parsedIP.BitLen()), nil
+		return maskedIPToLabel(ip, parsedIP.BitLen()), nil
 	} else {
 		parsedPrefix, err := netip.ParsePrefix(ip)
 		if err != nil {
 			return Label{}, fmt.Errorf("%q is not a CIDR: %w", ip, err)
 		}
-		return maskedIPToLabel(parsedPrefix.Masked().Addr(), parsedPrefix.Bits()), nil
+		return maskedIPToLabel(parsedPrefix.Masked().Addr().String(), parsedPrefix.Bits()), nil
 	}
 }
 
@@ -176,7 +174,7 @@ func computeCIDRLabels(cache *simplelru.LRU[netip.Prefix, []Label], lbls Labels,
 	}
 
 	// Compute the label for this prefix (e.g. "cidr:10.0.0.0/8")
-	prefixLabel := maskedIPToLabel(prefix.Addr(), ones)
+	prefixLabel := maskedIPToLabel(prefix.Addr().String(), ones)
 	lbls[prefixLabel.Key] = prefixLabel
 
 	// Keep computing the rest (e.g. "cidr:10.0.0.0/7", ...).